Cleanup summaries test, reorganize tests
