+++ x-pack/heartbeat/scenarios/basics_test.go
@@ -5,6 +5,7 @@
 package scenarios
 
 import (
+	"github.com/elastic/beats/v7/heartbeat/hbtestllext"
 	_ "github.com/elastic/beats/v7/heartbeat/monitors/active/http"
 	_ "github.com/elastic/beats/v7/heartbeat/monitors/active/icmp"
 	_ "github.com/elastic/beats/v7/heartbeat/monitors/active/tcp"
@@ -43,12 +44,9 @@ func TestSimpleScenariosBasicFields(t *testing.T) {
 	})
 }
 
-func TestLightweightSummaries(t *testing.T) {
+func TestLightweightUrls(t *testing.T) {
 	Scenarios.RunTag(t, "lightweight", func(mtr *MonitorTestRun, err error) {
-		require.GreaterOrEqual(t, len(mtr.Events()), 1)
-		lastCg := ""
-
-		for i, e := range mtr.Events() {
+		for _, e := range mtr.Events() {
 			testslike.Test(t, lookslike.MustCompile(map[string]interface{}{
 				"url": map[string]interface{}{
 					"full":   isdef.IsNonEmptyString,
@@ -56,16 +54,24 @@ func TestLightweightSummaries(t *testing.T) {
 					"scheme": mtr.StdFields.Type,
 				},
 			}), e.Fields)
+		}
+	})
+}
 
-			// Ensure that all check groups are equal and don't change
-			cg, err := e.GetValue("monitor.check_group")
-			require.NoError(t, err)
-			cgStr := cg.(string)
-			if i == 0 {
-				lastCg = cgStr
-			} else {
-				require.Equal(t, lastCg, cgStr)
-			}
+func TestLightweightSummaries(t *testing.T) {
+	Scenarios.RunTag(t, "lightweight", func(mtr *MonitorTestRun, err error) {
+		all := mtr.Events()
+		lastEvent, firstEvents := all[len(all)-1], all[:len(all)-1]
+		testslike.Test(t, lookslike.MustCompile(map[string]interface{}{
+			"summary": map[string]interface{}{
+				"up":   hbtestllext.IsUint16,
+				"down": hbtestllext.IsUint16,
+			},
+		}), lastEvent.Fields)
+
+		for _, e := range firstEvents {
+			summary, _ := e.GetValue("summary")
+			require.Nil(t, summary)
 		}
 	})
 }
